Simplify debug containers library to single header with basic allocation tracking
