+++ README.md
@@ -9,8 +9,7 @@ debug_containers/
 ├── include/
 │   └── memory/
 │       └── container_debug/
-│           ├── debug_containers.hpp    # Advanced header with context capture
-│           └── simple_debug_containers.hpp  # Simple non-intrusive header
+│           └── debug_containers.hpp    # Main header file
 ├── test/
 │   ├── basic/                          # Basic functionality tests
 │   │   ├── example.cpp
@@ -66,24 +65,8 @@ make
 ./test_ros_integration
 ```
 
-### Using the Headers
+### Using the Header
 
-#### Simple Non-Intrusive Approach (Recommended)
-```cpp
-#include <memory/container_debug/simple_debug_containers.hpp>
-
-int main() {
-    // Set memory threshold (default: 20MB)
-    Debug::set_memory_threshold(1024 * 1024); // 1MB
-    
-    // Use debug containers - automatically captures file/line/function
-    Debug::vector<int> my_vector(1000); // Will log with caller's location
-    
-    return 0;
-}
-```
-
-#### Advanced Context Capture Approach
 ```cpp
 #include <memory/container_debug/debug_containers.hpp>
 
@@ -91,12 +74,8 @@ int main() {
     // Set memory threshold (default: 20MB)
     Debug::set_memory_threshold(1024 * 1024); // 1MB
     
-    // Use debug containers with explicit context capture
-    Debug::vector<int> my_vector(__FILE__, __LINE__, __FUNCTION__);
-    
-    // Or use automatic context capture
-    AUTO_CONTEXT();
-    Debug::vector<int> my_vector2(1000); // Will log with caller's location
+    // Use debug containers - automatically captures file/line/function
+    Debug::vector<int> my_vector(1000); // Will log with caller's location
     
     return 0;
 }
@@ -110,7 +89,7 @@ int main() {
 - **ROS Integration**: Seamless integration with ROS logging
 - **Minimal Overhead**: Zero performance impact when allocations are below threshold
 - **Complete Coverage**: All major std containers supported
-- **Two Implementation Options**: Simple non-intrusive and advanced context capture approaches
+
 
 ## Available Containers
